backend/pyproject.toml
[tool.poetry]
name = "py-app-service"
version = "0.1.0"
description = ""
authors = ["hafid abi d <aselole87@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.112.2"
uvicorn = "^0.30.6"
motor = "^3.5.1"
pydantic = "^2.8.2"
web3 = "^7.5.0"
pyjwt = "^2.9.0"
black = "^24.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
